Transaction Functions to utils.js
