--- tests/test.js
@@ -1,5 +1,0 @@
-const ethNodes = require('./eth-nodes')
-const web3Scripts = require('./web3-scripts')
-
-// *TODO*
-// Run a test that tests each of the scripts in web3Scripts
